moved to 'startGame' handler to gameHandler.ts
