+++ src/server/handlers/communicationHandler.ts
@@ -14,15 +14,6 @@ export class CommunicationHandler implements HandlerInterface{
                 console.error("Couldn't deploy Message.")
             }
         });
-        socket.on('startGame', (data) => {  //TODO check data
-            let room = RoomHandler.getRoom(socket.id,lobbys);
-            if(room != undefined){
-                room.lobby.game = new Game(room.lobby.lobbyID,100,room.lobby.connections);
-                room.lobby.game.init(io);
-            }else{
-                console.error("room is undefined!");
-            }
-        });
     }
 
     /**
