+++ src/server/handlers/gameHandler.ts
@@ -3,6 +3,8 @@ import {Socket, Server as SocketServer} from "socket.io";
 import {LinkedList} from "typescriptcollectionsframework";
 import {GameLobby} from "../gameLobby";
 import {CommunicationHandler} from "./communicationHandler";
+import {RoomHandler} from "./roomHandler";
+import {Game} from "../game";
 
 export class GameHandler implements HandlerInterface {
 
@@ -25,6 +27,16 @@ export class GameHandler implements HandlerInterface {
         socket.on('canvasStatus',  (data) => {
             CommunicationHandler.deployMessage(socket, data, 'canvasUpdate', false, lobbys, io)
         })
+
+        socket.on('startGame', (data) => {  //TODO check data
+            let room = RoomHandler.getRoom(socket.id,lobbys);
+            if(room != undefined){
+                room.lobby.game = new Game(room.lobby.lobbyID,100,room.lobby.connections);
+                room.lobby.game.init(io);
+            }else{
+                console.error("room is undefined!");
+            }
+        });
     }
 }
 export let handler = new GameHandler();
